feat: add structure folder in project
